server/requirements.txt
eventlet==0.36.1
Flask==3.0.2
Flask-SocketIO==5.3.6
greenlet==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1